Refactor course in database
This was done to support moving questions to a
new course, add status for both question and
sesion, and have more control over the creation of a course

- All files and features has been changed to work with
new database design
- Changed course table into 5 tables (Course,
CourseCode, CourseSemester, Season, Year)
- Added check to stop user from creating session
or question before a course has been selected or
created.
- Added visual aid for admin to see that a course
isn't selected or if none is created.
- Added a complete overhaul for create new course
modal, with locale and validation both on client and
server for a new course.

- All features has been tested with the new overhaul and is working.