Added new feature
Any student assistant or admin, will now have a list on the admin dashboard listing all courses that they don't have access to. They can there apply to be an admin or an student assistant. An admin of that course will then have a list of applicants where they can either accept them or reject the application. There will also be a list with courses you have applied to