Completed feature
Admin and student assistant can now test their questions. Testing a question will open a modal showing the question. The user will be able to test the question as if it was the student answering. It will be possible to check the answer against the solution checker. This will give the result and the admin will be able to check their answer against the solution.

Adding a question to a session can now only be done to inactive sessions

Fixed bug preventing an admin from opening EditQuestionModal

Added locales for all new features