Changed how starting a session works
A host that have left a session does still have 15 minutes to rejoin a session before it will be removed from active sessions. If the host tries to start a new session, an error message will be displayed. The host will get the option to rejoin the session, or to remove the session from active sessions. If the host removes the session, all connected students will be moved to the session over screen

removed outdated TODO statements

Added locale for all new features